run forge fmt

+++ src/governance/PHOGovernorBravoDelegator.sol
@@ -4,39 +4,46 @@ pragma solidity ^0.8.10;
 import "./PHOGovernorBravoInterfaces.sol";
 
 contract PHOGovernorBravoDelegator is GovernorBravoDelegatorStorage, GovernorBravoEvents {
-	constructor(
-			address timelock_,
-			address pho_,
-			address admin_,
-	        address implementation_,
-	        uint votingPeriod_,
-	        uint votingDelay_,
-            uint proposalThreshold_) public {
-
+    constructor(
+        address timelock_,
+        address pho_,
+        address admin_,
+        address implementation_,
+        uint256 votingPeriod_,
+        uint256 votingDelay_,
+        uint256 proposalThreshold_
+    ) public {
         // Admin set to msg.sender for initialization
         admin = msg.sender;
 
-        // NOTE: if I'm understanding thisi correctly, implementation is being the delegate. It represents Delegator. Delegator encoded the `data`. `data` is as seen below. 
-        delegateTo(implementation_, abi.encodeWithSignature("initialize(address,address,uint256,uint256,uint256)",
-                                                            timelock_,
-                                                            pho_,
-                                                            votingPeriod_,
-                                                            votingDelay_,
-                                                            proposalThreshold_));
+        // NOTE: if I'm understanding thisi correctly, implementation is being the delegate. It represents Delegator. Delegator encoded the `data`. `data` is as seen below.
+        delegateTo(
+            implementation_,
+            abi.encodeWithSignature(
+                "initialize(address,address,uint256,uint256,uint256)",
+                timelock_,
+                pho_,
+                votingPeriod_,
+                votingDelay_,
+                proposalThreshold_
+            )
+        );
 
         _setImplementation(implementation_);
 
-		admin = admin_;
-	}
+        admin = admin_;
+    }
 
-
-	/**
+    /**
      * @notice Called by the admin to update the implementation of the delegator
      * @param implementation_ The address of the new implementation for delegation
      */
     function _setImplementation(address implementation_) public {
         require(msg.sender == admin, "GovernorBravoDelegator::_setImplementation: admin only");
-        require(implementation_ != address(0), "GovernorBravoDelegator::_setImplementation: invalid implementation address");
+        require(
+            implementation_ != address(0),
+            "GovernorBravoDelegator::_setImplementation: invalid implementation address"
+        );
 
         address oldImplementation = implementation;
         implementation = implementation_;
@@ -53,28 +60,26 @@ contract PHOGovernorBravoDelegator is GovernorBravoDelegatorStorage, GovernorBra
     function delegateTo(address callee, bytes memory data) internal {
         (bool success, bytes memory returnData) = callee.delegatecall(data);
         assembly {
-            if eq(success, 0) {
-                revert(add(returnData, 0x20), returndatasize())
-            }
+            if eq(success, 0) { revert(add(returnData, 0x20), returndatasize()) }
         }
     }
 
-	/**
+    /**
      * @dev Delegates execution to an implementation contract.
      * It returns to the external caller whatever the implementation returns
      * or forwards reverts.
      */
-    fallback () external payable {
+    fallback() external payable {
         // delegate all other functions to current implementation
-        (bool success, ) = implementation.delegatecall(msg.data);
+        (bool success,) = implementation.delegatecall(msg.data);
 
         assembly {
-              let free_mem_ptr := mload(0x40)
-              returndatacopy(free_mem_ptr, 0, returndatasize())
+            let free_mem_ptr := mload(0x40)
+            returndatacopy(free_mem_ptr, 0, returndatasize())
 
-              switch success
-              case 0 { revert(free_mem_ptr, returndatasize()) }
-              default { return(free_mem_ptr, returndatasize()) }
+            switch success
+            case 0 { revert(free_mem_ptr, returndatasize()) }
+            default { return(free_mem_ptr, returndatasize()) }
         }
     }
 }
